+++ packages/app/scripts/compile-language-extensions-into-json-file.ts
@@ -5,9 +5,9 @@ import { VSBuffer } from '@pkerschbaum/code-oss-file-service/out/vs/base/common/
 import { URI } from '@pkerschbaum/code-oss-file-service/out/vs/base/common/uri';
 import path from 'node:path';
 
+import type { LanguageExtensionPointJsonEntry } from '#pkg/base/file-icon-theme/load-css-rules-http';
 import { json as jsonUtil } from '#pkg/base/utils/json.util';
-import type { LanguageExtensionPointJsonEntry } from '#pkg/operations/file-icon-theme.operations';
-import { bootstrapDiskFileService } from '#pkg/platform/electron/file-explorer-agent/bootstrap-disk-file-service';
+import { bootstrapDiskFileService } from '#pkg/platform/electron/file-explorer-agent/disk-file-service';
 
 const CODE_OSS_FILE_ICON_THEME_LANGUAGES_PATH = URI.file(
   path.resolve(
