+++ packages/app/src/base/file-icon-theme/load-icon-classes.ts
