+++ packages/app/src/platform/electron/file-explorer-agent/disk-file-service.ts
@@ -4,13 +4,14 @@ import {
   LogService,
 } from '@pkerschbaum/code-oss-file-service/out/vs/platform/log/common/log';
 
+import type { IFileService } from '#pkg/base/files';
 import { FileService } from '#pkg/base/fileService';
 import { network } from '#pkg/base/network';
 
 export function bootstrapDiskFileService() {
   const logger = new ConsoleMainLogger();
   const logService = new LogService(logger);
-  const fileService = new FileService(logService);
+  const fileService: IFileService = new FileService(logService);
   const diskFileSystemProvider = new DiskFileSystemProvider(logService);
   fileService.registerProvider(network.Schemas.file, diskFileSystemProvider);
