+++ packages/app/src/platform/electron/file-icon-theme-loader.ts
@@ -1,4 +1,4 @@
-import { loadCssRules } from '#pkg/operations/file-icon-theme.operations';
+import { loadCssRules } from '#pkg/base/file-icon-theme/load-css-rules-http';
 import type { PlatformFileIconThemeLoader } from '#pkg/platform/file-icon-theme-loader.types';
 
 const FILE_ICON_THEME_RELATIVE_PATH = './icon-theme/';
